changed double to float and some other changes

+++ kart.h
@@ -1,10 +1,10 @@
 #ifndef KART_H
 #define KART_H
 
-#include <sdl2/SDL.h>
 #include "types.h"
+#include <sdl2/SDL.h>
 
-Kart *createKart(double x, double y, double rot, SDL_Texture *tex, double w);
+Kart *createKart(float x, float y, float rot, SDL_Texture *tex, float w);
 void updateKart(Kart *k);
 
 #endif
